requirements.txt
streamlit==1.41.1
pandas==2.2.3
plotly==5.24.1
numpy==2.2.1
pdfplumber==0.11.4
anthropic==0.42.0
python-dotenv==1.0.1
openpyxl==3.1.5
fpdf2==2.8.3
